+++ src/main/java/br/com/itau/geradornotafiscal/domain/frete/CalculoFrete.java
@@ -2,7 +2,9 @@ package br.com.itau.geradornotafiscal.domain.frete;
 
 import br.com.itau.geradornotafiscal.model.Pedido;
 
+import java.math.BigDecimal;
+
 public interface CalculoFrete {
 
-    double calcular(Pedido pedido);
+    BigDecimal calcular(Pedido pedido);
 }
